repo de calculo

+++ indexfile.cpp
@@ -20,7 +20,7 @@ Index_File::Index_File(){
 void Index_File::Add(Nodo node){
 }
 
-int Index_File::Find(long long int key){
+int Index_File::Find(long int key){
 
 }
 void Index_File::Reindex(){
